+++ apps/test-suite/tests/ExpoUIMeasurement.tsx
@@ -1,0 +1,7 @@
+export const name = 'ExpoUIMeasurement';
+
+export async function test({ describe, it }: any) {
+  describe(name, () => {
+    it('is only implemented on iOS', () => {});
+  });
+}